Move Section to components folder
