+++ src/components/Section/components/SectionDescription/index.tsx
