+++ src/components/Section/components/SectionSingleImage/index.tsx
