+++ src/components/Section/components/SectionText/index.tsx
