+++ src/containers/Features/index.tsx
@@ -1,7 +1,7 @@
-import SectionCardList from "../Section/components/SectionCardList";
-import SectionContentWrapper from "../Section/components/SectionContentWrapper";
-import SectionDescription from "../Section/components/SectionDescription";
-import { SectionCardType } from "../Section/components/Card";
+import SectionCardList from "../../components/Section/components/SectionCardList";
+import SectionContentWrapper from "../../components/Section/components/SectionContentWrapper";
+import SectionDescription from "../../components/Section/components/SectionDescription";
+import { SectionCardType } from "../../components/Section/components/Card";
 
 const cards: SectionCardType[] = [
   {
